build fclstm

+++ codes/fc_lstm.py
@@ -105,9 +105,9 @@ facet_train = torch.from_numpy(facet_train)
 covarep_train = covarep_train[:-valid_size]
 covarep_train = torch.from_numpy(covarep_train)
 y_valid = y_train[-valid_size:]
-y_valid = torch.from_numpy(y_valid)
+y_valid = torch.from_numpy(y_valid).float()
 y_train = y_train[:-valid_size]
-y_train = torch.from_numpy(y_train)
+y_train = torch.from_numpy(y_train).float()
 if args.cuda:
     embedding_train.cuda()
     embedding_valid.cuda()
